Yay! First fix I'm commiting on the way to 2.0.1 This is supposed to fix compilation problems of the curses module on SGI, and is based on the patch from 2.39 to 2.40 This should fix bug #116172
From the 2.39->2.40 log:
* Check for 'sgi' preprocessor symbol, not '__sgi__'
* Surround individual character macros with #ifdef's, instead of making them
  all rely on STRICT_SYSV_CURSES

Metacomment: this is as much to test procedures as it is
to actually fix the bug. If you're a CVS wizard please look
at the checkin message to see I'm commiting correctly,
and e-mail me if not, and I'll do my best to back it out and
reread the docos.

Good luck to me.

--HG--
branch : 2.0
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/release20-maint%4020144