Script to re-write @FOO@-style marks with values, initializing the replacement for @DATE@ from a TeX file containing a \date{...} mark (such as texinputs/boilerplate.tex).
This will be used to re-write the html/index.html.in file instead of
a combination of grep, date, and sed -- this is more portable to non-Unix
platforms.

This solves part of the problem reported in SF patch #429611, but does
not use the suggested patch.

--HG--
branch : legacy-trunk
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/legacy-trunk%4021835